fix: fix tarefaDAO, add initial validation

+++ app/src/main/java/br/ufpr/agenda/data/dao/TarefaDAO.kt
@@ -4,7 +4,6 @@ import android.content.ContentValues
 import android.content.Context
 import android.database.Cursor
 import br.ufpr.agenda.data.DBHelper
-import br.ufpr.agenda.data.model.StatusModel
 import br.ufpr.agenda.data.model.TarefaModel
 
 class TarefaDAO (private val context: Context) {
@@ -26,6 +25,11 @@ class TarefaDAO (private val context: Context) {
 
     fun update(tarefa: TarefaModel): Int {
         val db = dbHelper.writableDatabase
+
+        if (tarefa.id == 0) {
+            return 0
+        }
+
         val values = ContentValues().apply {
             put("titulo", tarefa.titulo)
             put("descricao", tarefa.descricao)
@@ -42,39 +46,11 @@ class TarefaDAO (private val context: Context) {
         return rowsAffected
     }
 
-    fun findByStatus(statusID: Int): List<TarefaModel> {
-        val db = dbHelper.readableDatabase
-        val cursor: Cursor = db.query(
-            DBHelper.STATUS_TABLE,
-            null,
-            "id_status=?",
-            arrayOf(statusID.toString()),
-            null,
-            null,
-            null
-        )
-
-        val tarefaList = mutableListOf<TarefaModel>()
-        while (cursor.moveToNext()) {
-            val status = TarefaModel(
-                id=cursor.getInt(cursor.getColumnIndexOrThrow("id")),
-                titulo=cursor.getString(cursor.getColumnIndexOrThrow("titulo")),
-                descricao=cursor.getString(cursor.getColumnIndexOrThrow("descricao")),
-                id_status=cursor.getInt(cursor.getColumnIndexOrThrow("id_status")),
-                id_prioridade=cursor.getInt(cursor.getColumnIndexOrThrow("id_prioridade"))
-            )
-            tarefaList.add(status)
-        }
-        cursor.close()
-        db.close()
-        return tarefaList
-    }
-
     fun find(statusId: Int? = null, prioridadeId: Int? = null, titulo: String? = null): List<TarefaModel> {
         val db = dbHelper.readableDatabase
         val conditions = mutableListOf<String>()
-        var args = mutableListOf<String>()
-        var where = ""
+        val args = mutableListOf<String>()
+        var where: String?
 
 
         if (titulo != null) {
@@ -93,13 +69,13 @@ class TarefaDAO (private val context: Context) {
         }
 
         if (!conditions.isEmpty()) {
-            var where = conditions.joinToString(" AND ")
+            where = conditions.joinToString(" AND ")
         } else {
-            var where = null
+            where = null
         }
 
         val cursor: Cursor = db.query(
-            DBHelper.STATUS_TABLE,
+            DBHelper.TAREFAS_TABLE,
             null,
             where,
             args.toTypedArray(),
@@ -110,42 +86,14 @@ class TarefaDAO (private val context: Context) {
 
         val tarefaList = mutableListOf<TarefaModel>()
         while (cursor.moveToNext()) {
-            val status = TarefaModel(
+            val tarefa = TarefaModel(
                 id=cursor.getInt(cursor.getColumnIndexOrThrow("id")),
                 titulo=cursor.getString(cursor.getColumnIndexOrThrow("titulo")),
                 descricao=cursor.getString(cursor.getColumnIndexOrThrow("descricao")),
                 id_status=cursor.getInt(cursor.getColumnIndexOrThrow("id_status")),
                 id_prioridade=cursor.getInt(cursor.getColumnIndexOrThrow("id_prioridade"))
             )
-            tarefaList.add(status)
-        }
-        cursor.close()
-        db.close()
-        return tarefaList
-    }
-
-    fun findAll(): List<TarefaModel> {
-        val db = dbHelper.readableDatabase
-        val cursor: Cursor = db.query(
-            DBHelper.STATUS_TABLE,
-            null,
-            null,
-            null,
-            null,
-            null,
-            null
-        )
-
-        val tarefaList = mutableListOf<TarefaModel>()
-        while (cursor.moveToNext()) {
-            val status = TarefaModel(
-                id=cursor.getInt(cursor.getColumnIndexOrThrow("id")),
-                titulo=cursor.getString(cursor.getColumnIndexOrThrow("titulo")),
-                descricao=cursor.getString(cursor.getColumnIndexOrThrow("descricao")),
-                id_status=cursor.getInt(cursor.getColumnIndexOrThrow("id_status")),
-                id_prioridade=cursor.getInt(cursor.getColumnIndexOrThrow("id_prioridade"))
-            )
-            tarefaList.add(status)
+            tarefaList.add(tarefa)
         }
         cursor.close()
         db.close()
@@ -154,6 +102,11 @@ class TarefaDAO (private val context: Context) {
 
     fun delete(id: Int): Int {
         val db = dbHelper.writableDatabase
+
+        if (id == 0) {
+            return 0
+        }
+
         val rowsDeleted = db.delete(
             DBHelper.TAREFAS_TABLE,
             "id=?",
